refactor(agent): propagate approval purpose through interactions
