+++ packages/agent/gui/agent-gui/agentGuiNode/AgentInteractivePromptSurface.spec.tsx
@@ -120,6 +120,7 @@ describe("AgentInteractivePromptSurface", () => {
           turnId: "turn-1",
           requestId: "request-file-change",
           callId: "request-file-change",
+          approvalPurpose: "edit-files",
           title: "Apply file changes",
           status: "waiting_approval",
           toolName: "Approval",
@@ -157,6 +158,37 @@ describe("AgentInteractivePromptSurface", () => {
     expect(screen.queryByText("Path")).toBeNull();
     expect(screen.getByText("Files")).toBeTruthy();
     expect(screen.getByText("app.js, styles.css")).toBeTruthy();
+  });
+
+  it("does not infer an edit-files title from file details", () => {
+    render(
+      <AgentInteractivePromptSurface
+        prompt={{
+          kind: "approval",
+          id: "approval:request-generic",
+          turnId: "turn-1",
+          requestId: "request-generic",
+          callId: "request-generic",
+          title: "Approval",
+          status: "waiting_approval",
+          toolName: "Approval",
+          input: {
+            changes: [{ path: "/workspace/session/app.js" }]
+          },
+          options: [{ id: "allow_once", label: "Allow", kind: "allow_once" }],
+          output: null,
+          occurredAtUnixMs: 1
+        }}
+        isSubmitting={false}
+        onSubmit={vi.fn()}
+        labels={labels}
+      />
+    );
+
+    expect(screen.getByText("Waiting for your approval")).toBeTruthy();
+    expect(
+      screen.queryByText("Codex requests authorization to edit files")
+    ).toBeNull();
   });
 
   it("clears approval option loading when external submission settles", async () => {
