+++ packages/agent/gui/shared/agentConversation/agentApprovalPurpose.ts
@@ -1,0 +1,7 @@
+export type AgentApprovalPurpose = "edit-files";
+
+export function normalizeAgentApprovalPurpose(
+  value: unknown
+): AgentApprovalPurpose | null {
+  return value === "edit-files" ? value : null;
+}
